+++ template/src/lib/react-query/query-client.ts
