+++ template/tsconfig.path.json
@@ -3,10 +3,10 @@
     "paths": {
       "@models/*": ["./src/model/*"],
       "@config": ["./config/env"],
-      "@redux-store/*": ["./src/redux/*"],
+      "@react-query/*": ["./src/lib/react-query/*"],
+      "@redux-store/*": ["./src/lib/redux/*"],
       "@lib/*": ["./src/lib/*"],
       "@i18n": ["./src/i18n"],
-      "@react-query/*": ["./src/react-query/*"],
       "@components-atoms/*": ["./src/components/atoms/*"],
       "@components-molecules/*": ["./src/components/molecules/*"],
       "@components-organisms/*": ["./src/components/organisms/*"],
